pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-csv-cleaner-env"
version = "0.1.0"
description = "CSV Data Cleaning Environment for OpenEnv - real-world data wrangling tasks for AI agents"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core] @ git+https://github.com/meta-pytorch/OpenEnv.git@v0.2.3",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "requests>=2.31.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "csv_cleaner_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["csv_cleaner_env", "csv_cleaner_env.server"]
package-dir = { "csv_cleaner_env" = ".", "csv_cleaner_env.server" = "server" }
